+++ tests/testthat/test-jzazbz100.R
@@ -2,9 +2,10 @@
 d65 <- whitepoints_cie1931("D65")
 
 # point chosen to be approx 50, 50, 50 in xyz100
+mat <- matrix(c(75.47733, 5.471956, 4.499812), ncol = 3)
 jab_test <-
   pth_new_jzazbz100(
-    matrix(c(75.47733, 5.471956, 4.499812), ncol = 3),
+    mat,
     whitepoint = d65
   )
 
@@ -29,5 +30,19 @@ test_that("pth_to_jzazbz100 works", {
   expect_equal(
     pth_to_jzazbz100(jab_test),
     jab_test
+  )
+})
+
+test_that("`[.pth_to_jzazbz100`() works", {
+
+  expect_identical(
+    jab_test,
+    jab_test[1, ]
+  )
+
+  expect_equal(
+    jab_test[ , 2:3],
+    mat[ , 2:3, drop = FALSE],
+    ignore_attr = TRUE
   )
 })